RPC: make RPCConnection non-internal, as it's a core API. Move docs around so they're on public API not internal API.
